refactors profile edit provider to provide profile as well
