+++ src/app/(main)/[username]/components/UserInfo.tsx
@@ -2,13 +2,12 @@ import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
 import { Skeleton } from '@/components/ui/skeleton';
 import DisplayName from '@/components/Profile/DisplayName';
 import Bio from '@/components/Profile/Bio';
-import { Profile } from '@/repositories/profileRepository/profile.types';
-import useProfile from '@/hooks/useProfile/useProfile';
 import SignOutBtn from '@/components/Buttons/SignOutBtn';
 import 'react-toastify/dist/ReactToastify.css';
+import { useProfileProvider } from '@/contexts/ProfileProvider';
 
-export default function UserInfo({ profile }: { profile: Profile }) {
-  const { isLoading } = useProfile(profile.id);
+export default function UserInfo() {
+  const { isLoading, profile } = useProfileProvider();
 
   return (
     <section className='flex flex-col gap-4 w-1/4 relative transform -translate-y-28'>
